+++ lib/isodoc/m3d/m3dbaserender.rb
@@ -1,8 +1,20 @@
+require_relative "metadata"
+require "fileutils"
+
 module IsoDoc
   module M3d
-    # A {Converter} implementation that generates CSAND output, and a document
-    # schema encapsulation of the document for validation
-    class PdfConvert < IsoDoc::PdfConvert
+    module BaseRender
+      def metadata_init(lang, script, labels)
+        @meta = Metadata.new(lang, script, labels)
+      end
+      
+      def add_image(filenames)
+        filenames.each do |filename|
+          FileUtils.cp html_doc_path(filename), File.join(@localdir, filename)
+          @files_to_delete << File.join(@localdir, filename)
+        end
+      end
+
       def annex_name(annex, name, div)
         div.h1 **{ class: "Annex" } do |t|
           t << "#{get_anchors[annex['id']][:label]} "
